Allows inheritance of callbacks
A child class can call .inherit_callbacks_before and .inherit_callbacks_after to
use the nearest ancestor's callbacks as their own. This can be useful when you
want to run callbacks defined in an ancestor class when the subclass's method is
called

Note that calling 'super' from an overridden instance method for which callbacks
have inherited from an ancestor, will continue to call the parent's callbacks
immediately before and/or after the call to 'super'.

Bumps to verion v0.2.0